register external using event
