+++ .gitignore
@@ -7,3 +7,4 @@
 vendor/
 composer.lock
 .phpunit.result.cache
+tests/storage
